Firstrun page basically working with new framework.
