+++ client/pages/firstrun.html
@@ -59,5 +59,6 @@
     <div id="donesection">
     <p><button id="done" type="submit"><img src="../assets/icon.png" /> Finished Setup</button></p>
     </div>
+    <script src="./firstrun.js"></script>
   </body>
 </html>
